% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdutilities.R
\docType{package}
\name{bdutilities}
\alias{bdutilities}
\alias{bdutilities-package}
\title{bdutilities: Reusable supporting functions across the bdverse.}
\description{
Internal functions required in bdverse package system.
}
\section{Citation}{

\itemize{
\item Gueta, T., Barve, V., Nagarajah, T., & Carmel, Y. (2019). bdutilities: Reusable supporting functions across the bdverse (R package V 0.0.1). Retrieved from https://github.com/bd-R/bdutilities/
}
}

