% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdverse_core_pkgs.R
\docType{data}
\name{bdverse_core_pkgs}
\alias{bdverse_core_pkgs}
\title{All \code{bdverse} core packages}
\format{
A character vector with eight elements
}
\usage{
bdverse_core_pkgs
}
\description{
All the names of \code{bdverse} \strong{core} packages (eight packages)
}
\keyword{datasets}
