% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{flatten_data}
\alias{flatten_data}
\title{Flatten data to be written to file}
\usage{
flatten_data(data)
}
\arguments{
\item{data}{dataframe, tibble, data.table}
}
\description{
Flatten data from multiple providers to make it suitable to be written to file.
}
