% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{return_core}
\alias{return_core}
\title{create a core object from reactive function}
\usage{
return_core(reactiveObject)
}
\arguments{
\item{reactiveObject}{shiny reactive object or, object}
}
\description{
get core object if its wrapped in shiny reactives.
}
