% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{summarize_dataframe}
\alias{summarize_dataframe}
\title{summarize biodiversity dataframe}
\usage{
summarize_dataframe(data)
}
\arguments{
\item{data}{biodiversity dataframe}
}
\description{
Summary table function.
}
