% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{summarizeDataframe}
\alias{summarizeDataframe}
\title{summarize biodiversity dataframe}
\usage{
summarizeDataframe(data)
}
\arguments{
\item{data}{biodiversity dataframe}
}
\description{
Summary table function.
}
