% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdverse_core_pkgs.R
\docType{data}
\name{bdverse_core_pkgs}
\alias{bdverse_core_pkgs}
\title{All \code{bdverse} core packages}
\format{
A data frame with 8 rows and 2 variables:
\describe{
\item{\code{pkg_name}}{Pakcage name}
\item{\code{pkg_color}}{Package color (Hex color code)}
}
}
\usage{
bdverse_core_pkgs
}
\description{
basic info  on all of \code{bdverse} \strong{core} packages (eight packages)
}
\keyword{datasets}
